+++ apps/dev/vite.config.js
@@ -1,4 +1,4 @@
-import { unding } from '@unding/vite';
+import { unding } from '@unding/studio/vite';
 
 export default {
   plugins: [unding()],
