+++ packages/unding/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@unding/vite",
+  "name": "@unding/studio",
   "version": "1.0.0",
   "description": "",
   "main": "src/index.js",
@@ -21,6 +21,7 @@
     "./pages/+onRenderHtml.js": "./src/pages/+onRenderHtml.js",
     "./pages/+onRenderClient.js": "./src/pages/+onRenderClient.js",
     "./pages/+config.js": "./src/pages/+config.js",
+    "./vite": "./src/vite/index.js",
     ".": "./src/index.js"
   },
   "bin": {
